+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support Rust's `stack-protector` feature (#1135)
 - Adding clock support for `ESP32-P4` (#1145)
 - Implementation OutputPin and InputPin for AnyPin (#1067)
+- Implement `estimate_xtal_frequency` for ESP32-C6 / ESP32-H2 (#1174)
 
 ### Fixed